add vue plugin
