+++ lib/config.js
@@ -1,32 +1,11 @@
-import Vue from 'vue' // optional
-
-const config = {
-  vue: Boolean(Vue),
-  vuexStore: null,
+export default {
   mutate(state, key, value) {
-    if (config.vue) {
-      const runMutation = () => {
-        if (key in state) {
-          state[key] = value
-        } else {
-          Vue.set(state, key, value)
-        }
-      }
-
-      if (config.vuexStore) {
-        config.vuexStore._withCommit(runMutation)
-      } else {
-        runMutation()
-      }
-    } else {
-      state[key] = value
-    }
+    state[key] = value
   },
+
   defaults: {
     fnIsEmpty: result => false, // eslint-disable-line no-unused-vars
     initialValue: null,
     spinnerDelay: 1000
   }
 }
-
-export default config
