+++ lib/index.js
@@ -5,5 +5,8 @@ import config from './config'
 export default {
   task: asyncTask,
   keyed: keyedAsyncTask,
+  use(newConfig) {
+    Object.assign(config, newConfig)
+  },
   config
 }
