+++ lib/keyed-async-task.js
@@ -1,7 +1,10 @@
 import helpers from './helpers'
 import config from './config'
 
-const m = config.mutate
+// Can't just do m = config.mutate, because config.mutate could change
+function m() {
+  config.mutate(...arguments)
+}
 
 /**
  * Creates and returns a keyed async task, with start() and clear() methods,
