+++ lib/plugins/vue.js
@@ -1,0 +1,22 @@
+import Vue from 'vue'
+
+export default function(store) {
+  return {
+    vuexStore: null,
+    mutate(state, key, value) {
+      const runMutation = () => {
+        if (key in state) {
+          state[key] = value
+        } else {
+          Vue.set(state, key, value)
+        }
+      }
+
+      if (store) {
+        store._withCommit(runMutation)
+      } else {
+        runMutation()
+      }
+    }
+  }
+}
